refactor: padronizar sobrecargas de ovld com typing.overload
- Adiciona blocos TYPE_CHECKING com @overload e pass em linha separada para Span.__init__ e Content.fit.
- Mantém o dispatch dinâmico do ovld sob o bloco else com assinaturas runtime compatíveis.
- Atualiza a sobrecarga de Layer.__init__ para aceitar tuple[float, float].

+++ src/anicrop/layer.py
@@ -76,7 +76,7 @@ class Layer(BaseLayer, AbstractLayer):
         @overload
         def __init__(
             self,
-            size: tuple[int, int],
+            size: tuple[float, float],
             *,
             opacity: float = 1.0,
             blend_mode: BlendMode = BlendMode.NORMAL,
